pyproject.toml
[project]
name = "commute-radio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.86.0",
    "edge-tts>=7.2.8",
    "fastmcp>=3.1.1",
    "httpx>=0.28.1",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
]
